change redirection logic to account for edge case with homepages
On the homepage of the Elder Scrolls wiki, there are several links
which point towards main pages for each Elder Scrolls game.  These
pages have titles of the form "Portal:[Elder Scrolls Game]".  In UESP,
the corresponding page has the title '[Game]:[Game]', and there is no
page with a title matching /^Portal:/ in all of the UESP wiki.

This commit is intended to address this issue by adding another phase
in the generation of the redirect link where, if the Wikia page name
matches the pattern /^Portal:(.+)$/, the user is redirected to the
page /\1:\1/, i.e. the UESP namespace is whatever the Wikia page is
called in the Wikia "Portal" namespace.