redirect skyrim.fandom.com pages to Skyrim namespace in UESP
This commit aims to address another major category of pages which the
original extension leaves unaltered: typically, pages on the Skyrim
wiki will belong to the "Skyrim:" namespace on the UESP wiki.

Doing this cleanly requires a significant refactor of the code as
previously written.  As a result, the namespace and page name are now
determined seperately, rather than all being found at once with
a single regex.  This has not caused any new bugs to emerge, at least
not in my testing.

Subsequent to this commit, two more changes are expected:
* Tests will be created to verify that my revisions have not created
  any major bugs
* One additional edge case will be accounted for --- the case when
  the user navigates directly to "elderscrolls.fandom.com" or
  "skyrim.fandom.com", which should redirect to the UESP home page
  and the Skyrim mainspace portal, respectively.