add testing to main branch and refactor to that end
This commit creates unit tests for the procedure for coming up with
a title for the redirected page.  redirect page title.  This required
changes in three distinct stages:

First, I refactored out this procedure into a separate function, now
located in the file `getUespPage.js`.

Second, because Firefox does not presently support modules in
extensions, this is packaged (along with `background-module.js`) into
the new background script at build time using esbuild (now registered
as a dev dependency).

Third, I added JestJS unit tests and corresponding test cases stored
in JSON format in the newly-created `__tests__` directory.  This also
necessitated the creation of a second node script, `node test`.  When
this script is run, test results are output to `testResults.txt`
(which is untracked).